requirements.txt
pyautogui
pydirectinput
pywinauto
requests
python-dotenv